+++ model/pdata/common_alias.go
@@ -23,10 +23,6 @@ import "go.opentelemetry.io/collector/pdata/pcommon"
 // Deprecated: [v0.49.0] Use pcommon.ValueType instead.
 type ValueType = pcommon.ValueType
 
-// AttributeValueType is an alias for pcommon.ValueType type.
-// Deprecated: [v0.48.0] Use ValueType instead.
-type AttributeValueType = pcommon.ValueType
-
 const (
 	// Deprecated: [v0.49.0] Use pcommon.ValueTypeEmpty instead.
 	ValueTypeEmpty = pcommon.ValueTypeEmpty
@@ -51,38 +47,11 @@ const (
 
 	// Deprecated: [v0.49.0] Use pcommon.ValueTypeBytes instead.
 	ValueTypeBytes = pcommon.ValueTypeBytes
-
-	// Deprecated: [v0.48.0] Use ValueTypeEmpty instead.
-	AttributeValueTypeEmpty = pcommon.ValueTypeEmpty
-
-	// Deprecated: [v0.48.0] Use ValueTypeString instead.
-	AttributeValueTypeString = pcommon.ValueTypeString
-
-	// Deprecated: [v0.48.0] Use ValueTypeInt instead.
-	AttributeValueTypeInt = pcommon.ValueTypeInt
-
-	// Deprecated: [v0.48.0] Use ValueTypeDouble instead.
-	AttributeValueTypeDouble = pcommon.ValueTypeDouble
-
-	// Deprecated: [v0.48.0] Use ValueTypeBool instead.
-	AttributeValueTypeBool = pcommon.ValueTypeBool
-
-	// Deprecated: [v0.48.0] Use ValueTypeMap instead.
-	AttributeValueTypeMap = pcommon.ValueTypeMap
-
-	// Deprecated: [v0.48.0] Use ValueTypeSlice instead.
-	AttributeValueTypeArray = pcommon.ValueTypeSlice
-
-	// Deprecated: [v0.48.0] Use ValueTypeBytes instead.
-	AttributeValueTypeBytes = pcommon.ValueTypeBytes
 )
 
 // Value is an alias for pcommon.Value struct.
 // Deprecated: [v0.49.0] Use pcommon.Value instead.
 type Value = pcommon.Value
-
-// Deprecated: [v0.48.0] Use Value instead.
-type AttributeValue = pcommon.Value
 
 // Aliases for functions to create pcommon.Value.
 var (
@@ -110,38 +79,11 @@ var (
 
 	// Deprecated: [v0.49.0] Use pcommon.NewValueBytes instead.
 	NewValueBytes = pcommon.NewValueBytes
-
-	// Deprecated: [v0.48.0] Use NewValueEmpty instead.
-	NewAttributeValueEmpty = pcommon.NewValueEmpty
-
-	// Deprecated: [v0.48.0] Use NewValueString instead.
-	NewAttributeValueString = pcommon.NewValueString
-
-	// Deprecated: [v0.48.0] Use NewValueInt instead.
-	NewAttributeValueInt = pcommon.NewValueInt
-
-	// Deprecated: [v0.48.0] Use NewValueDouble instead.
-	NewAttributeValueDouble = pcommon.NewValueDouble
-
-	// Deprecated: [v0.48.0] Use NewValueBool instead.
-	NewAttributeValueBool = pcommon.NewValueBool
-
-	// Deprecated: [v0.48.0] Use NewValueMap instead.
-	NewAttributeValueMap = pcommon.NewValueMap
-
-	// Deprecated: [v0.48.0] Use NewValueSlice instead.
-	NewAttributeValueArray = pcommon.NewValueSlice
-
-	// Deprecated: [v0.48.0] Use NewValueBytes instead.
-	NewAttributeValueBytes = pcommon.NewValueBytes
 )
 
 // Map is an alias for pcommon.Map struct.
 // Deprecated: [v0.49.0] Use pcommon.Map instead.
 type Map = pcommon.Map
-
-// Deprecated: [v0.48.0] Use Map instead.
-type AttributeMap = pcommon.Map
 
 // Aliases for functions to create pcommon.Map.
 var (
@@ -151,18 +93,3 @@ var (
 	// Deprecated: [v0.49.0] Use pcommon.NewMapFromRaw instead.
 	NewMapFromRaw = pcommon.NewMapFromRaw
 )
-
-// Deprecated: [v0.48.0] Use NewMap instead.
-var NewAttributeMap = pcommon.NewMap
-
-// Deprecated: [v0.48.0] Use Slice instead.
-type AttributeValueSlice = pcommon.Slice
-
-// Deprecated: [v0.48.0] Use NewSlice instead.
-var NewAttributeValueSlice = pcommon.NewSlice
-
-// Deprecated: [v0.48.0] Use InstrumentationScope instead.
-type InstrumentationLibrary = pcommon.InstrumentationScope
-
-// Deprecated: [v0.48.0] Use NewInstrumentationScope instead.
-var NewInstrumentationLibrary = pcommon.NewInstrumentationScope